pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Model Context Protocol (MCP) implementation for flight search and automated browser interactions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.29",
    "python-dotenv>=1.0.0",
    "mcp-use>=0.1.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=0.0.11"
]